modified su

+++ application/views/admin/category/index_view.php
@@ -35,7 +35,9 @@
                     echo '<td>'.$item->modified_date.'</td>';
                     echo '<td>';
                     echo '<a href="'.base_url("admin/").'category/edit/'.$item->id.'" style="color:#fff"><button class="btn btn-sm btn-warning">Sửa</button></a> ';
-                    echo '<a href="'.base_url("admin/").'category/delete/'.$item->id.'" style="color:#fff" onclick="return confirm(\'Bạn chắc chắn muốn xóa ?\')"><button class="btn btn-sm btn-danger">Xóa</button></a>';
+                    if($item->level != -1){
+                        echo '<a href="'.base_url("admin/").'category/delete/'.$item->id.'" style="color:#fff" onclick="return confirm(\'Bạn chắc chắn muốn xóa ?\')"><button class="btn btn-sm btn-danger">Xóa</button></a>';
+                    }                    
                     ?>
                     <?php
                     echo '</td>';
